(reviewed)
- compiler warnings and errors removed for msvc
- KHTMLWalletQueue moved to khtml_wallet_p.h (otherwise compilation wasn't possible)



svn path=/trunk/KDE/kdelibs/; revision=482263

+++ khtml/imload/pixmapplane.cpp
@@ -32,9 +32,9 @@ void PixmapPlane::paint(int dx, int dy, QPainter* p,
                   int sx, int sy, int sWidth, int sHeight)
 {
     //Do some basic clipping, discarding invalid requests and adjusting sizes of others
-    if (sy >= height)
+    if (sy >= (int)height)
         return;
-    if (sx >= width)
+    if (sx >= (int)width)
         return;
 
     if (sWidth == -1)
